Update Issue.cs
The proposed changes fix an issue if there are spaces in the key value.  Also, it allows the users to see all the keys.  The TrySetMember probably needs to be tweaked a little more to incorporate if someone tries to save the data and there are potential multiple items.  I could not get the system to access the object like an array but the PropertyByIndex_ works.